pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bugtriage-openenv"
version = "1.0.0"
description = "OpenEnv-compliant Bug Report Triage & Priority Queue Management environment"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.11.0",
    "requests>=2.32.3",
    "openai>=1.57.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "anyio>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "server*"]
